19-2-2026 10:10

+++ javascript/HuidigeToegang.js
@@ -1,8 +1,10 @@
 function BedrijfVerlaat() {
-    confirm("Weet je zeker dat de medewerker het bedrijf verlaat? \nDeze actie kan niet ongedaan worden gemaakt.");
+    return confirm("Weet je zeker dat de medewerker het bedrijf verlaat? \nDeze actie kan niet ongedaan worden gemaakt.");
 }
 
 function naamtoevoegen() {
-    const naam = document.getElementById("naamtoegangen").value;
-    localStorage.setItem("naamNieuweMedewerker", naam);
+    const naam = document.getElementById("naamtoegangen")?.value;
+    if (naam) {
+        localStorage.setItem("naamNieuweMedewerker", naam);
+    }
 }
